Fix CI lint: extract prunePreludeShadows into its own preprocessor module
The inline prune pass pushed typescriptPreprocessor.ts over the 1250
max-lines limit. Move it to lib/preprocessors/prunePreludeShadows.ts (a
standalone `prunePreludeShadows(program)` function, matching the pattern of
parallelDesugar / injectSchemaArgs / resolveReExports) with a co-located
test. No behavior change.
